pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-speech-min"
version = "0.1.0"
description = "Temporal threshold measurement and optimization for speech detection in Qwen models"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "OPRO Qwen Team"}
]

dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "soundfile>=0.12.1",
    "librosa>=0.10.1",
    # "webrtcvad>=2.0.10",  # Optional: requires C++ compiler on Windows
    "peft>=0.10.0",
    "accelerate>=0.27.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "pyannote.core>=5.0.0",
    "pyannote.database>=5.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "huggingface-hub>=0.21.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
vad = [
    "webrtcvad>=2.0.10",  # Requires Microsoft C++ Build Tools on Windows
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "E402"]  # E402: module level import not at top (needed for logging setup)

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
